tests/wpt/tests/css/css-fonts/parsing/font-synthesis-weight-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-synthesis-weight with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-synthesis-weight">
<meta name="assert" content="font-synthesis-weight supports the full grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-synthesis-weight', 'auto');
test_valid_value('font-synthesis-weight', 'none');
</script>
</body>
</html>
